refactor(aptu-coder): split exec_command, analyze_symbol, cache into sibling modules
Extract process-spawning logic from exec_command.rs into tools/exec_runtime.rs,
focused call-graph resolution from analyze_symbol.rs into tools/symbol_focused.rs,
and DiskCache impl from cache.rs into cache_disk.rs.

Each source file now falls below 600 LOC:
- exec_command.rs: 1075 -> 599 LOC
- analyze_symbol.rs: 1038 -> 566 LOC
- cache.rs: 976 -> 575 LOC

All existing public types and functions remain accessible at the same paths
via re-exports and module declarations.

Closes #1297
