+++ crates/aptu-coder/src/tests/filter_tests.rs
@@ -3,7 +3,7 @@ use regex::Regex;
 use crate::ShellOutput;
 use crate::filters;
 use crate::filters::{CompiledRule, apply_filter, maybe_inject_no_stat};
-use crate::tools::exec_command::handle_output_persist;
+use crate::tools::exec_runtime::handle_output_persist;
 use aptu_coder_core::types;
 
 #[test]
